add skywire cli log st ; move stCmd from cmd/skywire-cli/commands/rewards/ui.go
